Clena unused usings and sort usings

+++ mRemoteNG/UI/Controls/mrngIpTextBox.cs
@@ -3,11 +3,12 @@
  * Original Author: mawnkay
  */
 
-using System;
-using System.Windows.Forms;
-using mRemoteNG.Themes;
 using mRemoteNG.Resources.Language;
+using mRemoteNG.Themes;
+
+using System;
 using System.Runtime.Versioning;
+using System.Windows.Forms;
 
 namespace mRemoteNG.UI.Controls
 {
@@ -198,7 +199,7 @@ namespace mRemoteNG.UI.Controls
             this.Octet1.BackColor = System.Drawing.SystemColors.Menu;
             this.Octet1.BorderStyle = System.Windows.Forms.BorderStyle.None;
             this.Octet1.Font = new System.Drawing.Font("Segoe UI", 9F, System.Drawing.FontStyle.Regular,
-                                                       System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+                                                       System.Drawing.GraphicsUnit.Point, 0);
             this.Octet1.Location = new System.Drawing.Point(1, 1);
             this.Octet1.MaxLength = 3;
             this.Octet1.Name = "Octet1";
@@ -236,7 +237,7 @@ namespace mRemoteNG.UI.Controls
             // 
             this.Controls.Add(this.panel1);
             this.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular,
-                                                System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+                                                System.Drawing.GraphicsUnit.Point, 0);
             this.Name = "IPTextBox";
             this.Size = new System.Drawing.Size(124, 18);
             this.panel1.ResumeLayout(false);
