+++ mRemoteNG/UI/Controls/mrngRadioButton.cs
@@ -1,4 +1,5 @@
 ﻿using mRemoteNG.Themes;
+
 using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Runtime.Versioning;
@@ -132,7 +133,7 @@ namespace mRemoteNG.UI.Controls
             // NGRadioButton
             // 
             this.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular,
-                                                System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+                                                System.Drawing.GraphicsUnit.Point, 0);
             this.ResumeLayout(false);
         }
     }
