+++ mRemoteNG/UI/Window/BaseWindow.cs
@@ -1,6 +1,8 @@
 ﻿using mRemoteNG.Themes;
-using WeifenLuo.WinFormsUI.Docking;
+
 using System.Runtime.Versioning;
+
+using WeifenLuo.WinFormsUI.Docking;
 
 // ReSharper disable UnusedAutoPropertyAccessor.Global
 
@@ -71,7 +73,7 @@ namespace mRemoteNG.UI.Window
             // 
             this.ClientSize = new System.Drawing.Size(284, 261);
             this.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular,
-                                                System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+                                                System.Drawing.GraphicsUnit.Point, 0);
             this.Name = "BaseWindow";
             this.ResumeLayout(false);
         }
